[ST6RI-881] Refactored computation and caching of types of a Feature.
- Ensured that implicit types are computed when the types are derived.
